requirements.txt
aioredis==1.3.0
asgiref==3.2.2
asn1crypto==1.1.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.10.1
Automat==0.7.0
certifi==2019.9.11
cffi==1.13.0
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
Django==2.2.5
django-cors-headers==3.1.1
django-rest-swagger==2.2.0
djangorestframework==3.10.3
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
msgpack==0.6.2
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2-binary==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
PyQRCode==1.2.1
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.6
zope.interface==4.6.0
